backend/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-extensions==3.2.3
django-filter==23.2
djangorestframework==3.14.0
Markdown==3.4.3
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
